Refactor authorization boundaries and resource routing across Projects, Tasks, Workflows, and Project Members

+++ DevFlow.Application/Tasks/DeleteTask/DeleteTaskHandler.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using DevFlow.Application.Abstractions;
-using DevFlow.Application.Common.Interfaces;
+﻿using DevFlow.Application.Abstractions;
 using DevFlow.Application.Exceptions;
 using MediatR;
 
@@ -11,30 +7,36 @@ namespace DevFlow.Application.Tasks.DeleteTask
     public class DeleteTaskHandler : IRequestHandler<DeleteTaskCommand, DeleteTaskResult>
     {
         private readonly ITaskRepository _taskRepository;
-        private readonly IWorkspaceAuthorizationService _workspaceAuthorizationService;
         private readonly IUnitOfWork _unitOfWork;
-        private readonly ICurrentUserService _currentUserService;
-        public DeleteTaskHandler(ICurrentUserService currentUserService,IUnitOfWork unitOfWork, ITaskRepository taskRepository, IWorkspaceAuthorizationService workspaceAuthorizationService)
+        private readonly IProjectAuthorizationService _projectAuthorizationService;
+        public DeleteTaskHandler(
+        IUnitOfWork unitOfWork,
+        ITaskRepository taskRepository,
+        IProjectAuthorizationService projectAuthorizationService)
         {
-            _workspaceAuthorizationService = workspaceAuthorizationService;
             _unitOfWork = unitOfWork;
             _taskRepository = taskRepository;
-            _currentUserService=currentUserService;
+            _projectAuthorizationService = projectAuthorizationService;
         }
         public async Task<DeleteTaskResult> Handle(DeleteTaskCommand request, CancellationToken cancellationToken)
         {
-            var userId=_currentUserService.UserId;
-            var task = await _taskRepository.GetByIdForAdminAsync(request.TaskId,userId);
+            var task = await _taskRepository.GetByIdAsync(request.TaskId);
+
             if (task == null)
-            {
-                throw new NotFoundException("Task Doesnot Exist");
-            }
+                throw new NotFoundException("Task Does not exist.");
+
+            if (task.ProjectId != request.ProjectId)
+                throw new NotFoundException("Task Does not exist.");
+
+            await _projectAuthorizationService
+                .EnsureProjectMemberAsync(request.ProjectId);
+
             await _taskRepository.DeleteAsync(task);
             await _unitOfWork.SaveChangesAsync();
             var result = new DeleteTaskResult
             {
-                TaskId=task.Id,
-                Title=task.Title,
+                TaskId = task.Id,
+                Title = task.Title,
             };
             return result;
         }
